v5.x-remove-runtimecosting: sigma.util.Extensions moved to sigmastate.utils
